requirements.txt
pytorch
torchvision
# pytorch-cuda==12.1
opencv-python
pycocotools==2.0.7
matplotlib
numpy
seaborn==0.13.0
pandas
ultralytics
# transforms
# sgmllib3k
tensorboard
scikit-learn
tqdm
dlib
streamlit
